--- .commitmsg
@@ -1,1 +1,0 @@
-1594: AC.cpp (#1459) - AC,34.62%,69.23%
