+++ todo
@@ -1,5 +1,4 @@
-# rename Codes/1886-determine-whether-matrix-can-be-obtained-by-rotation.cpp.cpp
-# i == 3 break
+# i == 3 break |  Codes/1886-determine-whether-matrix-can-be-obtained-by-rotation.cpp
 
 # 记录iphone快捷指令 自动化 当到达某位置时触发
 
@@ -122,5 +121,3 @@ def get_environ_proxies(url, no_proxy=None):
         return getproxies()
     
 print(get_environ_proxies("https://letmefly.xyz"))
-
-# rm Codes\1594-maximum-non-negative-product-in-a-matrix.cpp
